Modularize redo-undo components and add icons
